# openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Agent Security Platform API
  version: 1.0.0
  description: Stable product API contract.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
paths:
...
  /api/v1/tests:
    get:
      operationId: listTests
      summary: List tenant-scoped red team definitions
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      parameters:
        - { name: cursor, in: query, required: false, schema: { type: string, minLength: 2, maxLength: 512, pattern: '^[A-Za-z0-9_-]+$' } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200': { description: Stable definition page., content: { application/json: { schema: { $ref: '#/components/schemas/TestDefinitionPage' } } } }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
    post:
      operationId: createTest
      summary: Create a bounded tenant-scoped red team definition
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      parameters:
        - $ref: '#/components/parameters/BrowserMutationCSRFToken'
        - $ref: '#/components/parameters/BrowserMutationOrigin'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ResourceVersion'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TestDefinitionInput' } } } }
      responses:
        '201':
          description: Durable definition.
          headers: { ETag: { $ref: '#/components/headers/WorkflowETag' }, X-Audit-ID: { $ref: '#/components/headers/WorkflowAuditID' }, X-Mutation-Receipt-ID: { $ref: '#/components/headers/WorkflowMutationReceiptID' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/TestDefinition' } } }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        '409': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
  /api/v1/tests/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/ProductID' } } ]
    get:
      operationId: getTest
      summary: Get one tenant-scoped red team definition
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      responses:
        '200': { description: Red team definition., headers: { ETag: { $ref: '#/components/headers/WorkflowETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/TestDefinition' } } } }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        '404': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
    patch:
      operationId: updateTest
      summary: Update one version-fenced red team definition
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      parameters:
        - $ref: '#/components/parameters/BrowserMutationCSRFToken'
        - $ref: '#/components/parameters/BrowserMutationOrigin'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ResourceVersion'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TestDefinitionUpdateInput' } } } }
      responses:
        '200': { description: Updated definition., headers: { ETag: { $ref: '#/components/headers/WorkflowETag' }, X-Audit-ID: { $ref: '#/components/headers/WorkflowAuditID' }, X-Mutation-Receipt-ID: { $ref: '#/components/headers/WorkflowMutationReceiptID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/TestDefinition' } } } }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        '404': { $ref: '#/components/responses/ProductErrorResponse' }
        '409': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
  /api/v1/tests/{id}/runs:
    parameters: [ { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/ProductID' } } ]
    post:
      operationId: runTest
      summary: Queue one durable red team run
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      parameters:
        - $ref: '#/components/parameters/BrowserMutationCSRFToken'
        - $ref: '#/components/parameters/BrowserMutationOrigin'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ResourceVersion'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TestRunInput' } } } }
      responses:
        '202': { description: Durable queued test run., headers: { ETag: { $ref: '#/components/headers/WorkflowETag' }, X-Audit-ID: { $ref: '#/components/headers/WorkflowAuditID' }, X-Mutation-Receipt-ID: { $ref: '#/components/headers/WorkflowMutationReceiptID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/TestRun' } } } }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        '404': { $ref: '#/components/responses/ProductErrorResponse' }
        '409': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
  /api/v1/test-runs:
    get:
      operationId: listTestRuns
      summary: List tenant-scoped red team runs
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      parameters:
        - { name: cursor, in: query, required: false, schema: { type: string, minLength: 2, maxLength: 1024, pattern: '^[A-Za-z0-9_-]+$' } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200': { description: Stable red team run page., content: { application/json: { schema: { $ref: '#/components/schemas/TestRunPage' } } } }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
  /api/v1/test-runs/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/ProductID' } } ]
    get:
      operationId: getTestRun
      summary: Get one redacted red team run and its immutable evidence attempts
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      responses:
        '200': { description: Redacted red team run detail., headers: { ETag: { $ref: '#/components/headers/WorkflowETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/TestRunDetail' } } } }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        '404': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
  /api/v1/test-runs/{id}/cancel:
    parameters: [ { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/ProductID' } } ]
    post:
      operationId: cancelTestRun
      summary: Request cancellation of one nonterminal red team run
      security:
        - BrowserSession: []
          BrowserExpectedScope: []
        - ProductAPIToken: []
      parameters:
        - $ref: '#/components/parameters/BrowserMutationCSRFToken'
        - $ref: '#/components/parameters/BrowserMutationOrigin'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ResourceVersion'
      responses:
        '200': { description: Cancelled or cancellation-requested run., headers: { ETag: { $ref: '#/components/headers/WorkflowETag' }, X-Audit-ID: { $ref: '#/components/headers/WorkflowAuditID' }, X-Mutation-Receipt-ID: { $ref: '#/components/headers/WorkflowMutationReceiptID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/TestRun' } } } }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        '401': { $ref: '#/components/responses/ProductErrorResponse' }
        '403': { $ref: '#/components/responses/ProductErrorResponse' }
        '404': { $ref: '#/components/responses/ProductErrorResponse' }
        '409': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
security:
  - BrowserSession: []
    BrowserExpectedScope: []
  - ProductAPIToken: []
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: __Host-zasp_session
      description: Secure, HttpOnly, SameSite=Lax host-only human browser session.
    BrowserExpectedScope:
      type: apiKey
      in: header
      name: X-Zasp-Expected-Scope
      description: Exact canonical organization/workspace/environment scope captured by the browser before a scoped request. The server compares it with the authenticated session scope before data access; ProductAPIToken scope remains token-owned.
    ProductAPIToken:
      type: http
      scheme: bearer
      bearerFormat: ProductAPIToken
      description: Product automation token.
  headers:
    WorkflowETag:
      description: Quoted durable resource version for optimistic concurrency.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    WorkflowAuditID:
      description: Durable audit record identifier for this mutation or its exact replay.
      schema: { $ref: '#/components/schemas/ProductID' }
    WorkflowMutationReceiptID:
      description: BrowserSession-only durable recovery receipt bound to the exact authenticated principal, scope, operation, intent, and attempt key. This header is absent for ProductAPIToken mutations and replays.
      schema: { $ref: '#/components/schemas/ProductID' }
...
  parameters:
...
    BrowserMutationCSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for BrowserSession mutations and omitted for ProductAPIToken mutations. The value is bound to the authenticated browser session.
      schema: { type: string, minLength: 32, maxLength: 256 }
    BrowserMutationOrigin:
      name: Origin
      in: header
      required: false
      description: Required for BrowserSession mutations and omitted for ProductAPIToken mutations. The server requires the exact configured same-origin HTTPS origin.
      schema: { type: string, minLength: 9, maxLength: 2048, pattern: '^https://[^/?#]+$' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated key binding an exact workflow mutation and its durable response.
      schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$' }
...
    ResourceVersion:
      name: If-Match
      in: header
      required: true
      description: Quoted current durable resource version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
...
  schemas:
...
    ProductID:
      type: string
      minLength: 40
      maxLength: 40
      pattern: '^pid_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      description: Canonical product identifier.
...
    TestDefinition:
      type: object
      additionalProperties: false
      required: [id, version, name, target_id, target_kind, categories, safety, enabled, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/ProductID' }
        version: { type: integer, minimum: 1, maximum: 1000000 }
        name: { type: string, minLength: 1, maxLength: 128 }
        target_id: { $ref: '#/components/schemas/ProductID' }
        target_kind: { type: string, enum: [agent_endpoint, mcp_server, coding_agent] }
        categories:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          items: { type: string, enum: [prompt_injection, tool_abuse, data_leakage, authorization_bypass, excessive_agency, sensitive_information] }
        safety: { $ref: '#/components/schemas/SafetyMetadata' }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TestDefinitionInput:
      type: object
      additionalProperties: false
      required: [id, name, target_id, target_kind, categories, safety]
      properties:
        id: { $ref: '#/components/schemas/ProductID' }
        name: { type: string, minLength: 1, maxLength: 128 }
        target_id: { $ref: '#/components/schemas/ProductID' }
        target_kind: { type: string, enum: [agent_endpoint, mcp_server, coding_agent] }
        categories: { type: array, minItems: 1, maxItems: 16, uniqueItems: true, items: { type: string, enum: [prompt_injection, tool_abuse, data_leakage, authorization_bypass, excessive_agency, sensitive_information] } }
        safety: { $ref: '#/components/schemas/SafetyMetadata' }
    TestDefinitionUpdateInput:
      type: object
      additionalProperties: false
      required: [name, target_id, target_kind, categories, safety, enabled]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        target_id: { $ref: '#/components/schemas/ProductID' }
        target_kind: { type: string, enum: [agent_endpoint, mcp_server, coding_agent] }
        categories: { type: array, minItems: 1, maxItems: 16, uniqueItems: true, items: { type: string, enum: [prompt_injection, tool_abuse, data_leakage, authorization_bypass, excessive_agency, sensitive_information] } }
        safety: { $ref: '#/components/schemas/SafetyMetadata' }
        enabled: { type: boolean }
    TestDefinitionPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/TestDefinition' } }
        next_cursor: { type: string, minLength: 2, maxLength: 512, pattern: '^[A-Za-z0-9_-]+$' }
    TestRunInput:
      type: object
      additionalProperties: false
      required: [run_id]
      properties:
        run_id: { $ref: '#/components/schemas/ProductID' }
    TestRun:
      type: object
      additionalProperties: false
      required: [id, version, definition_id, definition_version, status, attempt, cancel_requested, queued_at]
      properties:
        id: { $ref: '#/components/schemas/ProductID' }
        version: { type: integer, minimum: 1, maximum: 1000000 }
        definition_id: { $ref: '#/components/schemas/ProductID' }
        definition_version: { type: integer, minimum: 1, maximum: 1000000 }
        status: { type: string, enum: [queued, leased, retryable, complete, failed, cancelled] }
        attempt: { type: integer, minimum: 0, maximum: 5 }
        cancel_requested: { type: boolean }
        queued_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        verdict: { type: string, enum: [pass, fail, engine_error] }
        error_code: { type: string, enum: [retryable, rate_limited, denied, malformed, outcome_unknown, cancelled, exhausted] }
        evidence_reference: { type: string, minLength: 1, maxLength: 1024 }
...
    TestRunPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/TestRun' } }
        next_cursor: { type: string, minLength: 2, maxLength: 1024, pattern: '^[A-Za-z0-9_-]+$' }
...
    ProductError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - correlation_id
        - retryable
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z][a-z0-9]*(?:_[a-z0-9]+)*$'
        message:
          type: string
          minLength: 1
          maxLength: 512
          pattern: '^(?!\s)(?!.*\s$)[^\x00-\x1F\x7F-\x9F]+$'
        correlation_id:
          $ref: '#/components/schemas/ProductID'
        retryable:
          type: boolean
  responses:
    ProductErrorResponse:
      description: Request rejected with a product error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'